requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP Client
httpx==0.26.0

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# Database
sqlalchemy==2.0.25

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Task Scheduling
apscheduler>=3.10.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0

# Utilities
python-dotenv==1.0.1
fastapi-cache2==0.2.1
